P1T10

+++ pointers_arrays_strings/100-atoi.c
@@ -6,9 +6,9 @@
 */
 int _atoi(char *s)
 {
-int result = 0;
-int sign = 1;
-int i = 0;
+long int result = 0;
+long int sign = 1;
+long int i = 0;
 while (s[i] != '\0')
 {
 if (s[i] == '-')
